feat: logging location
i don't know how would this work yet, but right now i went with this:
log the error in the app layer, where it is met when the closure exits.
and the cli layer doesn't interact with the logger at all. yea, there
will be some repetition because the errors are not handled in one place,
but we have more control for the arguments instead, because we can log
the error with the command struct, result struct, handler name and so
on.